+++ ScientificReport/Controllers/OppositionController.cs
@@ -2,6 +2,8 @@ using System;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using ScientificReport.BLL.Interfaces;
+using ScientificReport.Controllers.Utils;
+using ScientificReport.DAL.Entities;
 using ScientificReport.DAL.Roles;
 using ScientificReport.DTO.Models.Opposition;
 
@@ -11,17 +13,25 @@ namespace ScientificReport.Controllers
 	public class OppositionController : Controller
 	{
 		private readonly IOppositionService _oppositionService;
+		private readonly IUserProfileService _userProfileService;
+		private readonly IDepartmentService _departmentService;
 
-		public OppositionController(IOppositionService oppositionService)
+		public OppositionController(
+			IOppositionService oppositionService,
+			IDepartmentService departmentService,
+			IUserProfileService userProfileService
+		)
 		{
 			_oppositionService = oppositionService;
+			_userProfileService = userProfileService;
+			_departmentService = departmentService;
 		}
 
 		// GET: Opposition
 		public IActionResult Index(OppositionIndexModel model)
 		{
-			model.Oppositions = _oppositionService.GetPage(model.CurrentPage, model.PageSize);
-			model.Count = _oppositionService.GetCount();
+			model.Oppositions = _oppositionService.GetPageByRole(model.CurrentPage, model.PageSize, User);
+			model.Count = _oppositionService.GetCountByRole(User);
 			return View(model);
 		}
 
@@ -39,14 +49,16 @@ namespace ScientificReport.Controllers
 				return NotFound();
 			}
 
+			if (!UserHasPermission(opposition))
+			{
+				return Forbid();
+			}
+
 			return View(opposition);
 		}
 
 		// GET: Opposition/Create
-		public IActionResult Create()
-		{
-			return View();
-		}
+		public IActionResult Create() => View();
 
 		// POST: Opposition/Create
 		[HttpPost]
@@ -57,6 +69,8 @@ namespace ScientificReport.Controllers
 			{
 				return View(model);
 			}
+
+			model.Opponent = _userProfileService.Get(User);
 			_oppositionService.CreateItem(model);
 			return RedirectToAction(nameof(Index));
 		}
@@ -74,6 +88,11 @@ namespace ScientificReport.Controllers
 			{
 				return NotFound();
 			}
+			
+			if (!UserHasPermission(opposition))
+			{
+				return Forbid();
+			}
 
 			return View(new OppositionEditModel(opposition));
 		}
@@ -88,6 +107,11 @@ namespace ScientificReport.Controllers
 				return NotFound();
 			}
 
+			if (!UserHasPermission(_oppositionService.GetById(id)))
+			{
+				return Forbid();
+			}
+			
 			if (!ModelState.IsValid)
 			{
 				return View(model);
@@ -110,6 +134,11 @@ namespace ScientificReport.Controllers
 			{
 				return NotFound();
 			}
+			
+			if (!UserHasPermission(opposition))
+			{
+				return Forbid();
+			}
 
 			return View(opposition);
 		}
@@ -124,8 +153,23 @@ namespace ScientificReport.Controllers
 				return NotFound();
 			}
 			
+			if (!UserHasPermission(_oppositionService.GetById(id)))
+			{
+				return Forbid();
+			}
+			
 			_oppositionService.DeleteById(id);
 			return RedirectToAction(nameof(Index));
+		}
+		
+		private bool UserHasPermission(Opposition opposition)
+		{
+			var user = _userProfileService.Get(User);
+			var department = _departmentService.Get(d => d.Staff.Contains(user));
+			return PageHelpers.IsAdmin(User) ||
+			       PageHelpers.IsHeadOfDepartment(User) &&
+			       department.Staff.Contains(opposition.Opponent) ||
+			       opposition.Opponent.Id == user.Id;
 		}
 	}
 }
